pyproject.toml
[project]
name = "chain-hello"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.8.3",
    "openai-agents>=0.4.0",
]
